Changed the visibility of helper methods in the FileLocator
These 2 methods are only used by locate() in the same file, so they
don't need to be public.
Thus, making a direct call to one of them would break the
resolution logic after a theme change as they always use the cached
theme name. The update of the theme based on the ActiveTheme is
done by locate only.